Tided up the make process for the libraries tboxlib and debuglib (debuglib was in with the same set of sources as tboxlib to I build it anyway).
All header files in the objects directory now get exported to
<Lib$Dir>.tboxlibs.h during export_hdrs phase. tboxlib and debuglib get build
with both zM and none zM flags and exported to <Lib$Dir>.tboxlibs.o.

Changed the sources slightly so they don't refer to the internal copy of the
header files found in objects but the copy previously exported to the right
place.