All hotlist_find_match calls wrapped in '#ifndef REMOTE_HOTLIST' compile time option, for Surfboard builds.
Cross referencing of images when the background colour is different - should
not do it for backgrounds, for e.g. alpha transparent PNGs would then fail.
This was implemented, but it turns out it affects *all* cross referencing!
Fixed for just background images.

As described in Docs.TagNames, the tag name "ahodgkin_127beta2" associated
with this checkin applies to everything *except* Choices.c. v1.27 beta 2
was build with the Choices file tagged "ahodgkin_127beta". The confusion
arises because Choices.c got independently updated and checked back in
before I checked in the v1.27 beta 2 changes myself.