Implemented size and age expiry of image history.  There are two ways of achieving this.
i)  There can be two sets of gadgets, one set for page history and another set
    for image history.

ii) There can be one set of gadgets for both settings and two radio buttons to
    select between them.

Only one of these may be present at the same time.  Currently the radio button
version is the one that is in place.  To change to the other version delete the
radio buttons and copy the gadgets from the ChoicesTMP window into the
appropriate window.