Enhanced animations support:
- Frame sequences are now explicitly specified.

- Frame durations are now explicitly specified, in tenths of a second.

- Animations can be aliased (e.g. a transmission animation can be
aliased as a reception animation with frame sequence numbers in
the reverse order).

As before, the animation details are read from the application's
messages file, to allow easy configuration.

Note the second change required a modification to the value of
Poll_Delay_Animations.  Users of UILib should either not be
sensitive to the Wimp PollIdle setting, or use the value of
Poll_Delay_Animations to calibrate themselves; they should not
assume any particular value (within reasonable limits!).

The code now assumes that all animations have been split up into
left, middle and right, where necessary, and that the sprites have
been appropriately named (e.g. for the "anim" animation, the
sprites should be named "anim_l#", "anim_#" and "anim_r#"
respectively, where # is the frame number).

This is version 0.29 of UILib; the previous version has been
tagged as version 0.28.