NFELib 0.38
Added a comment to !ChangeLog for NFELib.c++ (see NFELib.c++ for changelist)
0.38 also has some fixes in it for bugs that were found during the code
tidyup.  Daytona Serial number byte ordering was broken and the timesliced
FSMs could return NFE_INTERNALERR because they could get entered if the
timeslice expired before they were initialised.  The NFE_NOTENOUGHTTIME
code was set, but there was no check before entering the FSM loop to see
if the errorcode was clear.  This is now done.