Main changes from version 0.13:
- Tweaked transmit timings to fall more in line with T.30 (partial fix
for PAN-00754).

- Only allow BUSY and NO DIALTONE to be reported for transmission
(on reception only expect CONNECT).

- Add reporting of NO ANSWER and NO CARRIER, again only for transmission.
Note that if the T1 timeout (35 s) is smaller than the value of modems'
S-register S7 (typically 50 s by default) these will actually not be
reported (T1 timeout will occur first).

- Add informational return code Info_Connected, sent the first time
CONNECT is received from the modem (for both transmission and
reception).  Note that, being informational, this return code can
conceivably get squashed by subsequent return codes and so not be
sent, and so clients should not rely on it.

Tagged as version 0.14.