AcornPOP3 0.22
Logon fixed so that the USER and PASS commands don't send the null char to
the server.  getmsgfile.c fixed so that for GetMessage and GetXLinesOfMessage
the NO_SPACE error can be returned now.  The error returned to the client
was always getting zeroed before return, so the error was lost.

RESOLVERFAILED error moved in error enum to before NOSPACE error as a
duplicate NULLADDR was there and got removed.  However, this shifted all
the subsequent errors down by one. eg. NOSPACE became 17, before was 18.
This is now fixed so error numbers have their old correct values.

get FSM memory alloc size increased by one to make sure there is space for
null char should one get appended to the string.