Single character cookie names weren't being accepted as valid.  They now are.  This means that my.yahoo.com will now work properly (at least as far as cookie support is concerned, anyway ;-)
Also fixed a bug that was yet to appear but that would have hit eventually
whereby you couldn't load more than 100 cookies from a previously saved
cookie file.

Security errors are thrown if one of the secure SWIs is called and the SSL
services are not available any more.

Several comments added to explain deviances from standards and why we need
them.

More fixes to portlist handling to allow parameter-less Port attributes to
work on cookies.